Add scenic.forms framework
We have scenic.compat for supporting standard Django forms and
scenic.forms for forms designed in slightly more scenic manner. ie.
Create the classes when setting up the view and have them operate on the
request information storing all state on a per-request state object.

Very limited support at the moment and draws a line between the form
suppport and the widgets by deepcopying the form's widgets on every
request which is what Django does anyway in the BaseForm constructor.

This means that any code that treats widget instances as transient
per-request objects can continue to do so. This is what Django does
anyway as widgets are often associated with long lived field objects on
the Form classes as so aren't per-request anyway.