Convert list & single object views & adds tests
We remove the ObjectList helper as unless we do a strict evaluation of
the queryset, there really isn't anything to cache. We might add it back
in for exactly that reason at some point but for the moment we can
simply use a LiteralValue for the queryset to represent a list of
objects in the context.