Switch our main to use the template_view helper
So that we trigger the bug in the code with our tests and can fix it.

The template_view helper hadn't previously been updated to use the
TemplateResponse level of the hierarchy.

We also change the Makefile to set the PYTHONPATH so that we run the
local version of the scenic code rather than the installed version for
more immediate feedback.

And we add a helper to actually install it.