Store ObjectList value under named attr
We had it under a fixed attr name of 'object_list' before but that meant
you couldn't have multiple ObjectLists in the same view. Then we changed
it to store on the ObjectList instance itself but that is permanent for
the Django process and so caches the data and never changes even when
there are changes to the database.

Here we store on the 'state' but under a custom name to allow multiple
values on the same state.